backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
supabase==2.6.0
pydantic[email]